timing report

+++ OnlyT.Tests/TestTimingReport.cs
@@ -9,26 +9,48 @@
     public class TestTimingReport
     {
         private const int TotalMtgLengthMins = 105;
-        private const int SongAndPrayerMins = 10;
-
+        
         private readonly Random _random = new Random();
 
         [TestMethod]
-        public void TestWeekend()
+        public void TestReportGeneration()
         {
-            var startOfMtg = DateTime.Today + TimeSpan.FromHours(10);
-            var plannedEnd = startOfMtg.AddMinutes(TotalMtgLengthMins - SongAndPrayerMins);
+            var dateTimeService = new DateTimeServiceForTests();
 
-            DateTimeServiceForTests dateTimeService = new DateTimeServiceForTests();
-            dateTimeService.Set(startOfMtg);
+            const int weekCount = 20;
+            var dateOfFirstMeeting = GetNearestDayOnOrAfter(DateTime.Today.AddDays(-weekCount * 7), DayOfWeek.Sunday).Date;
+
+            for (int wk = 0; wk < weekCount; ++wk)
+            {
+                var dateOfWeekendMtg = dateOfFirstMeeting.AddDays(wk * 7);
+                var dateOfMidweekMtg = dateOfWeekendMtg.AddDays(4);
+
+                StoreWeekendData(wk, dateOfWeekendMtg, dateTimeService);
+                StoreMidweekData(wk, weekCount, dateOfMidweekMtg, dateTimeService);
+            }
+        }
+
+        private void StoreWeekendData(
+            int week,
+            DateTime dateOfWeekendMtg, 
+            DateTimeServiceForTests dateTimeService)
+        {
+            var startOfMtg = dateOfWeekendMtg + TimeSpan.FromHours(10);
+            var plannedEnd = startOfMtg.AddMinutes(TotalMtgLengthMins);
 
             var service = new LocalTimingDataStoreService(null, dateTimeService);
-            service.DeleteAllData();
-            
+
+            if (week == 0)
+            {
+                service.DeleteAllData();
+            }
+
+            dateTimeService.Set(startOfMtg);
+
             service.InsertMeetingStart(startOfMtg);
             service.InsertPlannedMeetingEnd(plannedEnd);
 
-            dateTimeService.Add(TimeSpan.FromMinutes(1));
+            dateTimeService.Add(TimeSpan.FromSeconds(1));
 
             // song and prayer
             InsertTimer(service, dateTimeService, "Introductory Segment", false, 5);
@@ -47,9 +69,64 @@
             service.InsertActualMeetingEnd();
 
             service.Save();
+        }
 
-            var file = TimingReportGeneration.ExecuteAsync(service, null).Result;
-            Assert.IsNotNull(file);
+        private void StoreMidweekData(
+            int week,
+            int weekCount,
+            DateTime dateOfMidweekMtg,
+            DateTimeServiceForTests dateTimeService)
+        {
+            var startOfMtg = dateOfMidweekMtg + TimeSpan.FromHours(19);
+            var plannedEnd = startOfMtg.AddMinutes(TotalMtgLengthMins);
+
+            var service = new LocalTimingDataStoreService(null, dateTimeService);
+
+            dateTimeService.Set(startOfMtg);
+
+            service.InsertMeetingStart(startOfMtg);
+            service.InsertPlannedMeetingEnd(plannedEnd);
+
+            dateTimeService.Add(TimeSpan.FromMinutes(1));
+
+            InsertTimer(service, dateTimeService, "Introductory Segment", false, 5);
+
+            InsertTimer(service, dateTimeService, "Opening Comments", false, 3);
+
+            InsertTimer(service, dateTimeService, "Treasures", false, 10);
+
+            InsertTimer(service, dateTimeService, "Digging for Spiritual Gems", false, 8);
+
+            InsertTimer(service, dateTimeService, "Bible Reading", true, 10);
+            dateTimeService.Add(TimeSpan.FromSeconds(70));
+
+            InsertTimer(service, dateTimeService, "Ministry Talk 1", false, 4);
+
+            InsertTimer(service, dateTimeService, "Ministry Talk 2", true, 3);
+            dateTimeService.Add(TimeSpan.FromSeconds(70));
+
+            InsertTimer(service, dateTimeService, "Ministry Talk 3", true, 6);
+            dateTimeService.Add(TimeSpan.FromSeconds(70));
+
+            InsertTimer(service, dateTimeService, "Interim Segment", false, 5);
+
+            InsertTimer(service, dateTimeService, "Living Item 1", false, 15);
+
+            InsertTimer(service, dateTimeService, "Congregation Bible Study", false, 30);
+
+            InsertTimer(service, dateTimeService, "Review", false, 3);
+
+            InsertTimer(service, dateTimeService, "Concluding Segment", false, 5);
+
+            service.InsertActualMeetingEnd();
+
+            service.Save();
+
+            if (week == weekCount - 1)
+            {
+                var file = TimingReportGeneration.ExecuteAsync(service, null).Result;
+                Assert.IsNotNull(file);
+            }
         }
 
         private void InsertTimer(
@@ -62,6 +139,14 @@
             service.InsertTimerStart(talkDescription, isStudentTalk, TimeSpan.FromMinutes(targetMins), TimeSpan.FromMinutes(targetMins));
             dateTimeService.Add(GetDuration(targetMins));
             service.InsertTimerStop();
+
+            // add an amount for speaker swap
+            dateTimeService.Add(GetDuration(0.3));
+        }
+
+        private DateTime GetNearestDayOnOrAfter(DateTime dt, DayOfWeek dayOfWeek)
+        {
+            return dt.AddDays(((int)dayOfWeek - (int)dt.DayOfWeek + 7) % 7);
         }
 
         private TimeSpan GetDuration(double targetMinutes)
